move features.{cc,h} to brave/components
