+++ browser/sources.gni
@@ -136,7 +136,6 @@ brave_chrome_browser_deps = [
   "//brave/browser/misc_metrics",
   "//brave/browser/misc_metrics:misc_metrics_impl",
   "//brave/browser/ntp_background",
-  "//brave/browser/resources/settings/email_aliases_page:features",
   "//brave/browser/skus",
   "//brave/browser/sync",
   "//brave/browser/themes",
@@ -195,6 +194,7 @@ brave_chrome_browser_deps = [
   "//brave/components/debounce/content/browser",
   "//brave/components/decentralized_dns/content",
   "//brave/components/decentralized_dns/core",
+  "//brave/components/email_aliases:features",
   "//brave/components/google_sign_in_permission",
   "//brave/components/https_upgrade_exceptions/browser",
   "//brave/components/ipfs",
