+++ browser/ui/webui/settings/brave_privacy_handler.cc
@@ -9,13 +9,13 @@
 #include "base/functional/bind.h"
 #include "base/values.h"
 #include "brave/browser/brave_browser_features.h"
-#include "brave/browser/resources/settings/email_aliases_page/features.h"
 #include "brave/components/ai_chat/core/browser/utils.h"
 #include "brave/components/ai_chat/core/common/features.h"
 #include "brave/components/brave_shields/core/common/features.h"
 #include "brave/components/constants/pref_names.h"
 #include "brave/components/de_amp/common/features.h"
 #include "brave/components/debounce/core/common/features.h"
+#include "brave/components/email_aliases/features.h"
 #include "brave/components/google_sign_in_permission/google_sign_in_permission_util.h"
 #include "brave/components/p3a/pref_names.h"
 #include "brave/components/request_otr/common/buildflags/buildflags.h"
