+++ components/email_aliases/features.cc
@@ -3,7 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "brave/browser/resources/settings/email_aliases_page/features.h"
+#include "brave/components/email_aliases/features.h"
 
 namespace email_aliases {
